+++ app/src/main/java/edu/phystech/weather/database/forecast/hourly/entity/HourlyForecastDBEntity.kt
@@ -16,7 +16,7 @@ data class HourlyForecastDBEntity (
     @ColumnInfo(name = "feels_like") val feels_like: Float,
 
     @ColumnInfo(name = "humidity") val humidity: Long,
-    @ColumnInfo(name = "wind_speed") val wind_speed: Long,
+    @ColumnInfo(name = "wind_speed") val wind_speed: Float,
     @ColumnInfo(name = "wind_deg") val wind_deg: Long,
     @ColumnInfo(name = "uvi") val uvi: Float,
     @ColumnInfo(name = "icon") val icon: String,
